+++ src/nmk_workspace/resolvers.py
@@ -18,14 +18,18 @@ class SubProjectsResolver(NmkListConfigResolver):
     (other behaviors may be implemented if needed later).
     """
 
-    def get_value(self, name: str, root: str, only_nmk_projects: bool = True) -> list[str]:  # type: ignore
+    def get_value(self, name: str, root: str, only_nmk_projects: bool = True, only_python_projects: bool = False) -> list[str]:  # type: ignore
         """
         Resolver for sub-projects list.
 
         :param root: root path of the workspace
         :param only_nmk_projects: if True (default), only return sub-projects having a default nmk project file (nmk.yml)
+        :param only_python_projects: if True, only return sub-projects having a Python project file (pyproject.toml)
         :return: list of sub-projects paths relative to the workspace root
         """
+
+        # Build expected files list
+        expected_files: list[str] = [] + (["nmk.yml"] if only_nmk_projects else []) + (["pyproject.toml"] if only_python_projects else [])  # type: ignore
 
         # Ask git for submodules paths
         root_path = Path(root)
@@ -36,15 +40,17 @@ class SubProjectsResolver(NmkListConfigResolver):
             lambda x: x.group(1) if x is not None else None,
             filter(lambda x: x is not None, map(lambda x: SUB_MODULE_PATTERN.match(x.strip()), cp.stdout.splitlines(keepends=False))),
         ):
-            # Only keep ones with a default nmk model file
+            # Check candidate path is valid and exists
             assert candidate is not None  # for type hinting
             candidate_path = Path(candidate)
             if candidate_path.is_absolute() or not (root_path / candidate_path).exists():  # pragma: no cover
                 NmkLogger.debug(f"Sub-project path {candidate_path} is not valid, skipping it.")
                 continue
+
+            # Check for expected files, if any
             sub_module_path = candidate_path.as_posix()
-            if (not only_nmk_projects) or (root_path / candidate_path / "nmk.yml").is_file():
+            if (not expected_files) or all((root_path / candidate_path / file).is_file() for file in expected_files):
                 nmk_models.append(sub_module_path)
             else:
-                NmkLogger.debug(f"Sub-project {sub_module_path} does not have a default nmk model file, skipping it.")
+                NmkLogger.debug(f"Sub-project {sub_module_path} does not have some of expected files ({', '.join(expected_files)}), skipping it.")
         return nmk_models