+++ doc/changelog.md
@@ -11,6 +11,8 @@ The fully detailed changelog is also available on [Github](https://github.com/dy
 ## Release 1.3.0
 
 - Updated documentation theme
+- New **{ref}`${workspacePythonSubProjects}<workspacePythonSubProjects>`** config item listing all python subprojects.
+- Added {ref}`uv workspace members<uv-contrib>` generation in workspace-level **`pyproject.toml`** file.
 
 ## Release 1.2.0